+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-01-03 17:56:04
+
+### Changed Files
+- `supabase/functions/instagram-refresh-token/index.ts`
+
+### Statistics
+```
+ 1 file changed, 273 insertions(+)
+```
+
 ## [Checkpoint] - 2026-01-03 17:55:25
 
 ### Changed Files